# backend/core/schemas/bill_actions_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bill – Make Actions",
    "version": "1.3.1"
  },
  "servers": [
    {
      "url": "https://hook.eu2.make.com"
    }
  ],
  "paths": {
...
    "/w8kw2yvyhvt2zh9lf5m8gkus4slywdjq": {
  "post": {
    "operationId": "load_client_context",
    "summary": "Load full client context bundle",
    "description": "Returns the full client context JSON bundle (profile, plan context, weeks/sessions/steps, and related context). Input contract is a single top-level client_id.",
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "required": ["client_id"],
            "properties": {
              "client_id": {
                "type": "string",
                "description": "Client identifier (matched case-insensitively in Clients sheet)."
              }
            },
            "additionalProperties": false
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Client context loaded successfully",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["status", "client_id", "timestamp"],
              "properties": {
                "status": { "type": "string", "example": "OK" },
                "client_id": { "type": "string" },
                "timestamp": { "type": "string", "description": "Timestamp string generated by scenario." }
              },
              "additionalProperties": true
            }
          }
        }
      }
    }
  }
},
...
    "/kgujcny6vi04gng5rwawv4j4jljdhiff": {
      "post": {
        "operationId": "authenticate_developer",
        "summary": "Authenticate developer for Tech Mode",
        "description": "Validates a provided developer access key via Make scenario. Returns authenticated true/false. The key must not be logged or echoed.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["provided_key"],
                "properties": {
                  "provided_key": {
                    "type": "string",
                    "description": "Developer access key supplied by the user."
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["authenticated"],
                  "properties": {
                    "authenticated": {
                      "type": "boolean"
                    },
                    "status": {
                      "type": "string"
                    },
                    "action": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  },
                                   "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
  }
}